+++ CHANGELOG.md
@@ -5,7 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 * Get endpoint for cache expiration
-* Put endpoint for cache expiration
+* Put/Post endpoint for cache expiration
 
 ### Changed
 * bumped version of leaflet/esri leaflet used by preview